feat: simplify encoding config
